requirements.txt
Flask==2.3.0
pdfplumber==0.11.9
Pillow==12.1.1
matplotlib==3.10.8
Werkzeug==2.3.0
gunicorn==21.2.0
